@@ -1,11 +1,6 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
     `./src/pages/**/*.{js,jsx,ts,tsx}`,
     `./src/components/**/*.{js,jsx,ts,tsx}`,
   ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
 }
